pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "meeting-scheduler"
version = "0.1.0"
description = "A minimalistic and basic example multi-turn communication with an AI agent to schedule a meeting"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT License"}
authors = [
    {name = "Saibaba Telukunta", email = "saibaba.telukunta@gmail.com"},
]
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "pydantic",
    "python-dotenv",
    "langchain",
    "langchain-openai",
    "langgraph",
    "dateparser",
    "pytz",
]

[project.optional-dependencies]
dev = [
    "pytest", # Example dev dependency
]

